requirements.txt
streamlit==1.40.1
mysql.connector